+++ Package.swift
@@ -10,19 +10,19 @@ let package = Package(
     dependencies: [
         .package(
             url: "https://github.com/BroadApps-official/broad-extensions-ios.git",
-            exact: "1.0.0"
+            exact: "1.0.1"
         ),
         .package(
             url: "https://github.com/BroadApps-official/broad-core-ios.git",
-            exact: "1.0.0"
+            exact: "1.1.0"
         ),
         .package(
             url: "https://github.com/BroadApps-official/broad-monetization-ios.git",
-            exact: "1.0.0"
+            exact: "1.1.0"
         ),
         .package(
             url: "https://github.com/BroadApps-official/broad-ui-flows-ios.git",
-            exact: "1.0.0"
+            exact: "1.0.1"
         ),
         .package(
             url: "https://github.com/Swinject/Swinject.git",